requirements.txt
edge-tts==6.1.12
pydub==0.25.1
moviepy==1.0.3
Pillow==10.4.0
pyyaml==6.0.2
numpy==2.0.1